refactor(frontend): decouple productivity logic into ViewModel
Extract state, memos, and callbacks from `ProductivityScreen` into `useProductivityViewModel` to eliminate the "God Class" pattern. Extract `ProductivityListItem` as a memoized component. Existing logic and behaviors remain strictly intact.
